requirements.txt
langchain>=0.1.12
langchain-community>=0.0.28
langchain-core>=0.1.30
python-dotenv>=1.0.0
requests>=2.31.0
google-api-python-client>=2.118.0 